chore(verify): extract provider and better error
