+++ cli/src/cmd/forge/verify/sourcify.rs
@@ -1,9 +1,10 @@
-use super::{VerificationProvider, VerifyArgs, VerifyCheckArgs};
+use super::{VerifyArgs, VerifyCheckArgs};
 use crate::cmd::LoadConfig;
 use async_trait::async_trait;
 use cast::SimpleCast;
 use ethers::solc::ConfigurableContractArtifact;
 
+use crate::cmd::forge::verify::provider::VerificationProvider;
 use foundry_common::fs;
 use foundry_utils::Retry;
 use futures::FutureExt;
